Physical systems

Here we define the solvers of the physical systems.

Generalities

Methods that are common to several solvers.

WavKinS.Delta — Method

Delta(k::Float64, k1::Float64, k2::Float64)

Assuming that $k_1 < k + k_2$, $k_2 < k + k_1$, and $k < k_1 + k_2$, it returns twice the area of the triangle of sides $k$, $k_1$, and $k_2$, i.e. $\Delta = \frac{1}{2} \sqrt{(k + k_1 + k_2) (k + k_1 - k_2) (k - k_1 + k_2) (-k + k_1 + k_2)} = 2 \left[ (k k_1)^2 + (k k_2)^2 + (k_1 k_2)^2 \right] - k^4 - k_1^4 - k_2^4$.

WavKinS.Δ_vs — Method

Δ_vs(k::Float64, p::Float64, q::Float64)

Return $\Delta = \frac{1}{2} \sqrt{(k^2-p^2) q (2 k+q)}$. It is Delta with $k_1 = (k + p + q)/2$ and $k_2 = (k - p + q)/2$.

WavKinS.θkmax — Method

θkmax(k,kmax)

Gives ones if $k < k_{max}$, $0$, otherwise.